only build fuzzers for go1.18+
