+++ tree_test.go
@@ -990,20 +990,3 @@ func TestInsertResolveSplitLeaf(t *testing.T) {
 		t.Fatal("didn't find the resolved leaf where expected")
 	}
 }
-
-func FuzzStatelessVsStateful(f *testing.F) {
-	f.Add([]byte{})
-	f.Fuzz(func(t *testing.T, input []byte) {
-		rootF := New()
-		rootL := NewStateless()
-
-		for i := 0; i < len(input)/64; i++ {
-			rootF.Insert(input[i*64:i*64+32], input[i*64+32:(i+1)*64], nil)
-			rootL.Insert(input[i*64:i*64+32], input[i*64+32:(i+1)*64], nil)
-		}
-
-		if !Equal(rootL.ComputeCommitment(), rootF.ComputeCommitment()) {
-			t.Fatalf("root commitment for state-less != -ful %x != %x", rootF.ComputeCommitment().Bytes(), rootL.ComputeCommitment().Bytes())
-		}
-	})
-}
